新增根据图片url删除图片接口和完善文件上传模块

+++ src/modules/upload/upload.service.ts
@@ -6,12 +6,13 @@ import * as crypto from 'crypto';
 import { PrismaClientKnownRequestError } from '@prisma/client/runtime/library';
 import * as fs from 'fs';
 import * as mime from 'mime-types';
+import { IUploadEntity } from './upload.entity';
 
 @Injectable({ scope: Scope.REQUEST })
 export class UploadService {
   constructor(@Inject(REQUEST) private request: Request,private readonly prisma: PrismaService) {}
 
-  async uploadFile(file,type): Promise<string> {
+  async uploadFile(file): Promise<string> {
     const fileName = Date.now() + extname(file.originalname);
     try {
       // 检查文件类型是否为图片
@@ -55,32 +56,37 @@ export class UploadService {
     }
   }
 
-  async deleteFile(id: number): Promise<boolean> {
+  async deleteFile(url,id: number = 0,): Promise<boolean> {
     try {
+      const checkParams = (id === 0 && url === "") || (id !== 0 && url !== "")
       // 从数据库中获取文件记录
-      const fileRecord = await this.prisma.image.findUnique({
-        where: { id },
-      });
-
-      if (!fileRecord) {
-        throw new Error('文件记录不存在');
+      if (checkParams) {
+        throw new HttpException('请传入正确的参数',HttpStatus.OK)
       }
-
+      // 从数据库中获取文件记录
+      const fileRecord: IUploadEntity = await this.prisma.image.findFirst({
+        where: { OR: [{ id }, { url } ] },
+      });
+      if (!fileRecord) {
+        throw new HttpException('文件记录不存在',HttpStatus.OK);
+      }
+      if (fileRecord.num > 0) {
+        throw new HttpException('文件存在引用',HttpStatus.OK);
+      }
       // 删除文件系统中的文件
       const filePath = fileRecord.url.split('/').slice(-2).join('/');
-      console.log(filePath);
       fs.unlink(`./${filePath}`, async (err) => {
-          if (err) {
-            console.error('删除文件时发生错误:', err);
-            throw new Error('删除文件失败');
-          } else {
-            // 从数据库中删除记录
-            await this.prisma.image.delete({
-              where: { id },
-            });
-            return true;
-          }
-        });
+        if (err) {
+          console.error('删除文件时发生错误:', err);
+          throw new HttpException('删除文件失败',HttpStatus.INTERNAL_SERVER_ERROR);
+        } else {
+          // 从数据库中删除记录
+          await this.prisma.image.deleteMany({
+            where: { OR: [{ id }, { url } ] },
+          });
+          return true;
+        }
+      });
     } catch (error) {
       if (error instanceof PrismaClientKnownRequestError) {
         throw new HttpException('删除失败', HttpStatus.INTERNAL_SERVER_ERROR);
